requirements.txt
fastapi==0.115.12
gunicorn==23.0.0
uvicorn==0.34.3
SQLAlchemy==2.0.41
google-cloud-bigquery
google-cloud-storage